requirements.txt
matplotlib~=3.10.7
numpy~=2.3.4
pytransform3d~=3.14.3
pyserial~=3.5
open3d==0.19.0